Fix wordpress installation.
Our wordpress setup steps were failing for several reasons including the
fact that gawk wasn't installed on the Docker container. I simplified
the setup a bit so that we don't need to parse the wp-config file at
all. Instead, we prepend our own db definitions to the top of the
wp-config.php using a require to our own probo-config.php file.