Allow the wp-cli to run as root.
Since we run as root in the docker container, wp-cli throws errors when
running. This is meant to help you avoid doing silly things on real
servers but since we are running on a container, it is ok. Prevent these
warnings and allow the script to continue by supplying the --allow-root
flag.