Add Drupal version config option and validation.
We want to change the behavior of the builds based on the version of
Drupal that is being built. We don't necessarily know this, or at least
don't want to bother parsing all the files to figure it out so we allow
the user to specify it in the .probo.yaml file. The default version is
currently set to 7, with 6 and 8 set to the min and max versions
respectivly.