Wait as long as needed for MySQL to start.
This resloves an issue in which, under high server load, MySQL
server takes a long time to start up. In these cases, builds hand
forever because the default behavior was to wait for a set period
of time and then begin running build steps. Often in these cases
the build would fail for lack of a MySQL connection. Instead, we
can use netcat and hold us in a loop until netcat is able to
reach MySQL on the standard port, only then allowing the build
steps to continue. After 60 seconds this will fail exiting 1.