Capture reason for reap.
We want to know why the builds have been reaped. Here we allow
the delete endpoint to recieve query parameters for the reason.
This should take an integer code. The codes are defined by the
reaper but they should be non-negative integers. The reason is
then stored along with the build data in the database.