Create pin/unpin endpoints.
This change allows put requests to be made for builds which will
set them to pinned or unpinned. Updating a build result in the
new status for the build being stored in the db (level) and also
will result in a build event being sent. The reaper can capture
this build event as needed.