Fix wordpress database name.
Our settings allow the user to specify the name of the database
that is used in the application, however we were not properly
creating the WP settings. Without this change, if a user
specifies the database name, the config file will have the proper
values and the application will not properly connect to the
database.