Adjust WP options and docs.
This change adjusts the docs for the wordpress plugin and renames
the devDomain and devHome options to wpDomain and wpHome. This is
intended to reduce confusion as many users will be providing the
live site database and code so calling it "dev" doesn't make much
sense. The other changes fix the database replacement of these
values if they are not set so that we don't accidently do search
and replace in the database for strings such as "null" or
"undefined". It also provides an option to update plugins at the
time that the build is running.