Add a 'running' status; Use globally defined status constants
This update creates a new status / state called 'running,' to define the state when a task or build is actually in progress. As a result, in the future, only a task or build that has been scheduled but not started will be in the 'pending' state. Additionally, this update forces the use of status constants, so that there is no variation in the status naming convention (e.g. a failed build is always in the 'error' state, not 'failed' or 'failure').