Update package list before attempting to install.
We allow users to specify additional packages they would like
installed in the probo.yaml file but we don't actually update the
package list. As a result the package can often not be found and
the task will fail. This change ensures that if additional
packages are requested, we update the apt list before attempting
to install.